requirements.txt
fastapi
uvicorn
torch
pillow
pandas
git+https://github.com/openai/CLIP.git 